fix: use int for stock and quantity, remove unused BigDecimal import

+++ src/main/java/org/example/pooexamfinal/dto/StockResponse.java
@@ -1,10 +1,8 @@
 package org.example.pooexamfinal.dto;
 
-import java.math.BigDecimal;
-
 public record StockResponse(
         String id,
         String name,
-        long stock
+        int stock
 ) {
 }
